+++ src/views/map/sources/ActiveUAVsLayerSource.js
@@ -13,6 +13,7 @@ import { source, withLayer } from '@collmot/ol-react';
 import FeatureManager from '../FeatureManager';
 import UAVFeature from '../features/UAVFeature';
 
+import { errorCodeToSemantics } from '~/flockwave/errors';
 import Flock from '~/model/flock';
 import { setLayerSelectable, setLayerTriggersTooltip } from '~/model/layers';
 import { uavIdToGlobalId } from '~/model/identifiers';
@@ -229,12 +230,15 @@ class ActiveUAVsLayerSource extends React.Component {
         uav.lon,
         uav.lat,
       ]);
+      const mostSevereError = uav.mostSevereError;
 
       // Set or update the heading of the feature
       if (uav.heading !== undefined) {
         feature.heading = uav.heading;
       }
 
+      feature.status =
+        mostSevereError > 0 ? errorCodeToSemantics(mostSevereError) : null;
       feature.labelColor = this.props.labelColor;
     }
   };